+++ envoy-control-core/src/main/kotlin/pl/allegro/tech/servicemesh/envoycontrol/groups/NodeMetadataValidator.kt
@@ -1,13 +1,11 @@
 package pl.allegro.tech.servicemesh.envoycontrol.groups
 
 import io.envoyproxy.controlplane.server.DiscoveryServerCallbacks
-import io.envoyproxy.envoy.api.v2.DiscoveryResponse
 import pl.allegro.tech.servicemesh.envoycontrol.groups.CommunicationMode.ADS
 import pl.allegro.tech.servicemesh.envoycontrol.groups.CommunicationMode.XDS
 import pl.allegro.tech.servicemesh.envoycontrol.logger
 import pl.allegro.tech.servicemesh.envoycontrol.protocol.HttpMethod
 import pl.allegro.tech.servicemesh.envoycontrol.snapshot.SnapshotProperties
-import io.envoyproxy.envoy.api.v2.DiscoveryRequest as DiscoveryRequestV2
 import io.envoyproxy.envoy.config.core.v3.Node as NodeV3
 import io.envoyproxy.envoy.service.discovery.v3.DiscoveryRequest as DiscoveryRequestV3
 
@@ -65,27 +63,12 @@ class NodeMetadataValidator(
         request?.node?.let { validateV3Metadata(it) }
     }
 
-    override fun onV2StreamRequest(streamId: Long, request: DiscoveryRequestV2?) {
-        request?.node?.let { validateV2Metadata() }
-    }
-
-    override fun onStreamResponse(
-        streamId: Long,
-        request: DiscoveryRequestV2?,
-        response: DiscoveryResponse?
-    ) {
-    }
-
     private fun validateV3Metadata(node: NodeV3) {
         // Some validation logic is executed when NodeMetadata is created.
         // This may throw NodeMetadataValidationException
         val metadata = NodeMetadata(node.metadata, properties)
 
         validateMetadata(metadata)
-    }
-
-    private fun validateV2Metadata() {
-        throw V2NotSupportedException()
     }
 
     private fun validateMetadata(metadata: NodeMetadata) {