+++ envoy-control-core/src/main/kotlin/pl/allegro/tech/servicemesh/envoycontrol/server/callbacks/MetricsDiscoveryServerCallbacks.kt
@@ -2,7 +2,6 @@ package pl.allegro.tech.servicemesh.envoycontrol.server.callbacks
 
 import io.envoyproxy.controlplane.cache.Resources
 import io.envoyproxy.controlplane.server.DiscoveryServerCallbacks
-import io.envoyproxy.envoy.api.v2.DiscoveryRequest as V2DiscoveryRequest
 import io.envoyproxy.envoy.service.discovery.v3.DiscoveryRequest as V3DiscoveryRequest
 import io.micrometer.core.instrument.MeterRegistry
 import java.util.concurrent.atomic.AtomicInteger
@@ -18,11 +17,6 @@ class MetricsDiscoveryServerCallbacks(private val meterRegistry: MeterRegistry) 
         companion object {
             fun fromTypeUrl(typeUrl: String) = when (typeUrl) {
                 // TODO_deprecate_v2: do we need this mapping still?
-                Resources.V2.CLUSTER_TYPE_URL -> CDS
-                Resources.V2.ENDPOINT_TYPE_URL -> EDS
-                Resources.V2.LISTENER_TYPE_URL -> LDS
-                Resources.V2.ROUTE_TYPE_URL -> RDS
-                Resources.V2.SECRET_TYPE_URL -> SDS
                 Resources.V3.CLUSTER_TYPE_URL -> CDS
                 Resources.V3.ENDPOINT_TYPE_URL -> EDS
                 Resources.V3.LISTENER_TYPE_URL -> LDS
@@ -56,11 +50,6 @@ class MetricsDiscoveryServerCallbacks(private val meterRegistry: MeterRegistry) 
     }
 
     override fun onV3StreamRequest(streamId: Long, request: V3DiscoveryRequest) {
-        meterRegistry.counter("grpc.requests.${StreamType.fromTypeUrl(request.typeUrl).name.toLowerCase()}")
-            .increment()
-    }
-
-    override fun onV2StreamRequest(streamId: Long, request: V2DiscoveryRequest) {
         meterRegistry.counter("grpc.requests.${StreamType.fromTypeUrl(request.typeUrl).name.toLowerCase()}")
             .increment()
     }