+++ CHANGELOG.md
@@ -43,6 +43,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Access log filters config handled
 - Global snapshot auditor added
 - updated project java version 11
+- updated java-control-plane to version with removed api v2
 
 ## [0.19.5]
 ### Changed